split `api` into `configuration` and `thread_pool` modules
